Visualization preview implementation. Move Dataset preview button to TimeFilters widget. Helper service added.
